fix(gcloud): fixup cloudrun traffic split on deploy
`model-api` is currently experiencing TOOLING-2T, which turns out to
have been stemmed by the fact that our live revision is from January --
new deployments have not actually been getting any traffic.

This changeset ensures that after every deployment we ensure that the
new deployment is the one to which we feed traffic.