feat(linter): allow executor overrides
This allows users to override the entire executor, rather than just the
python version and the resource class. By allowing this option, users
should be able to have more customization options such as attaching
environment variables to their image, running background images for
testing purposes, etc.

Note that this is a breaking change!