feat(linter): allow args override
Users may want to override the arguments provided to pre-commit.

For instance, in my case we are introducing these checks in large
existing repositories, and fixing pre-commit issues for all files
upfront is impractical (and would potentially break things).

Until we fix all issues in the tested repository, we'll run with
`--from-ref=origin/main --to-ref=HEAD` as to test files that have been
updated in the PR.