fix(linter): support PEP668
With PEP668, global package installations are no longer possible, even
in something like docker where an additional layer of virtualenvs is
entirely meaningless. As such, we need to overhaul this orb to install
pre-commit in a venv, which in turn means alpine-based executors need
shell/env overrides and a single command for "install and run
pre-commit" is not really doable anymore: there are too many different
possible base images to account for them all.